+++ src/main/java/org/opentripplanner/api/resource/PlannerResource.java
@@ -77,7 +77,7 @@ public class PlannerResource extends RoutingResource {
             res = routingService.route(request, router);
 
             // Map to API
-            TripPlanMapper tripPlanMapper = new TripPlanMapper(request.locale);
+            TripPlanMapper tripPlanMapper = new TripPlanMapper(request.locale, request.showIntermediateStops);
             response.setPlan(tripPlanMapper.mapTripPlan(res.getTripPlan()));
             response.setMetadata(TripSearchMetadataMapper.mapTripSearchMetadata(res.getMetadata()));
             if (!res.getRoutingErrors().isEmpty()) {