+++ src/main/java/org/opentripplanner/model/FlexStopLocation.java
@@ -1,7 +1,5 @@
 package org.opentripplanner.model;
 
-import java.util.Collection;
-import java.util.List;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Point;
 
@@ -16,6 +14,8 @@ public class FlexStopLocation extends TransitEntity implements StopLocation {
   private String name;
 
   private Geometry geometry;
+
+  private String zoneId;
 
   public FlexStopLocation(FeedScopedId id) {
     super(id);
@@ -52,5 +52,14 @@ public class FlexStopLocation extends TransitEntity implements StopLocation {
   public WgsCoordinate getCoordinate() {
     Point centroid = geometry.getCentroid();
     return new WgsCoordinate(centroid.getY(), centroid.getX());
+  }
+
+  @Override
+  public String getFirstZoneAsString() {
+    return zoneId;
+  }
+
+  public void setZoneId(String zoneId) {
+    this.zoneId = zoneId;
   }
 }