+++ src/main/java/org/opentripplanner/routing/fares/impl/MultipleFareServiceFactory.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.fares;
+package org.opentripplanner.routing.fares.impl;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -6,9 +6,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.opentripplanner.model.OtpTransitService;
-import org.opentripplanner.routing.impl.DefaultFareServiceFactory;
-import org.opentripplanner.routing.services.FareService;
-import org.opentripplanner.routing.services.FareServiceFactory;
+import org.opentripplanner.routing.fares.FareService;
+import org.opentripplanner.routing.fares.FareServiceFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;