+++ src/main/java/org/opentripplanner/routing/fares/impl/Ride.java
@@ -1,10 +1,13 @@
-package org.opentripplanner.routing.impl;
+package org.opentripplanner.routing.fares.impl;
 
+import java.time.LocalTime;
+import java.time.ZonedDateTime;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Stop;
 
 import java.util.HashSet;
 import java.util.Set;
+import org.opentripplanner.model.StopLocation;
 
 /**
  * A set of edges on a single route, with associated information. Used only in calculating fares.
@@ -23,17 +26,17 @@ public class Ride {
 
     String endZone;
 
-    long startTime;
+    ZonedDateTime startTime;
 
-    long endTime;
+    ZonedDateTime endTime;
 
     // in DefaultFareServiceImpl classifier is just the TraverseMode
     // it can be used differently in custom fare services
     public Object classifier;
 
-    public Stop firstStop;
+    public StopLocation firstStop;
 
-    public Stop lastStop;
+    public StopLocation lastStop;
 
     public Ride() {
         zones = new HashSet<String>();